[Feature] エラー処理 #9
- erros/に各エラーの基本となるメッセージを定義
- middleware/error.jsにエラーに対応するミドルウェア関数を定義
- controllers/にエラー処理を反映
- ルーティングが見つからないページへの対応

+++ client/controllers/injury.js
@@ -2,28 +2,28 @@ const Injury = require("../models/injury");
 const Transfer = require("../models/transfer");
 const Player = require("../models/player");
 const Team = require("../models/team");
+const { StatusCodes } = require("http-status-codes");
+const { NotFoundError, BadRequestError } = require("../errors");
 
 const getAllInjury = async (req, res) => {
-  const injurys = await Injury.find({});
-  res.status(200).json({ data: injurys });
+  const injuries = await Injury.find({});
+  res.status(StatusCodes.OK).json({ data: injuries });
 };
 
 const createInjury = async (req, res) => {
+  if (!req.body.team || !req.body.player) {
+    throw new BadRequestError();
+  }
   const { team, player } = req.body;
 
   const playerObj = await Player.findOne({ abbr: player });
   if (!playerObj) {
-    return res
-      .status(404)
-      .json({ message: "登録されていない選手が選択されています" });
+    throw new NotFoundError();
   }
 
   const teamObj = await Team.findOne({ abbr: team });
-
   if (!teamObj) {
-    return res
-      .status(404)
-      .json({ message: "登録されていないチームが選択されています" });
+    throw new NotFoundError();
   }
 
   const now_teamObj = await Transfer.findOne({
@@ -41,18 +41,30 @@ const createInjury = async (req, res) => {
   };
 
   const injury = await Injury.create(injuryData);
-  res.status(200).json({ message: "追加しました", data: injury });
+  res
+    .status(StatusCodes.CREATED)
+    .json({ message: "追加しました", data: injury });
 };
 
 const getInjury = async (req, res) => {
+  if (!req.params.id) {
+    throw new BadRequestError();
+  }
   const {
     params: { id: injuryId },
   } = req;
   const injury = await Injury.findById(injuryId);
-  res.status(200).json({ data: injury });
+  if (!injury) {
+    throw new NotFoundError();
+  }
+
+  res.status(StatusCodes.OK).json({ data: injury });
 };
 
 const updateInjury = async (req, res) => {
+  if (!req.params.id || !req.body.team || !req.body.player) {
+    throw new BadRequestError();
+  }
   const {
     params: { id: injuryId },
     body,
@@ -64,7 +76,7 @@ const updateInjury = async (req, res) => {
   if (body.team && !mongoose.Types.ObjectId.isValid(body.team)) {
     const teamObj = await Team.findOne({ abbr: body.team });
     if (!teamObj) {
-      return res.status(404).json({ message: "team が見つかりません" });
+      throw new BadRequestError("team が見つかりません。");
     }
     updatedData.team = teamObj._id;
   }
@@ -73,7 +85,7 @@ const updateInjury = async (req, res) => {
   if (body.player && !mongoose.Types.ObjectId.isValid(body.player)) {
     const playerObj = await Player.findOne({ abbr: body.player });
     if (!playerObj) {
-      return res.status(404).json({ message: "player が見つかりません" });
+      throw new BadRequestError("player が見つかりません。");
     }
     updatedData.player = playerObj._id;
   }
@@ -86,18 +98,27 @@ const updateInjury = async (req, res) => {
       runValidators: true,
     }
   );
+  if (!injury) {
+    throw new NotFoundError();
+  }
 
-  res.status(200).json({ message: "編集しました" });
+  res.status(StatusCodes.OK).json({ message: "編集しました" });
 };
 
 const deleteInjury = async (req, res) => {
+  if (!req.params.id) {
+    throw new BadRequestError();
+  }
   const {
     params: { id: injuryId },
   } = req;
 
   const injury = await Injury.findOneAndDelete({ _id: injuryId });
+  if (!injury) {
+    throw new NotFoundError();
+  }
 
-  res.status(200).json({ message: "削除しました" });
+  res.status(StatusCodes.OK).json({ message: "削除しました" });
 };
 
 module.exports = {
